Refactor: Split useParticipations hook into smaller files
This commit refactors the useParticipations hook into smaller, more manageable files to improve code organization and readability.
